microflow/pyproject.toml
[project]
name = "microflow"
version = "0.1.0"
description = ""
authors = [
    {name = "watacoso",email = "watacoso@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "google-cloud-storage (>=3.1.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "microflow", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
responses = "^0.25.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
